Move macros defines to STL and refactor base.hpp and STL headers Macro defines moved to STL to make those available in it and ensure consistency across the codebase
