+++ Engine/Runtime/include/core/internal/config.h
@@ -1,7 +1,7 @@
 #pragma once
 
-#define RK_CRT_CHECKS_ENABLED 1
-#define RK_ASSERTIONS_ENABLED 1
-#define RK_EXCEPTIONS_ENABLED 1
-#define RK_LOG_DEBUG_ENABLED  1
-#define RK_LOG_TRACE_ENABLED  1
+#define RKSTL_CRT_CHECKS_ENABLED 1
+#define RKSTL_ASSERTIONS_ENABLED 1
+#define RK_EXCEPTIONS_ENABLED    1
+#define RK_LOG_DEBUG_ENABLED     1
+#define RK_LOG_TRACE_ENABLED     1
