+++ Engine/Runtime/include/metadata.hpp
@@ -2,10 +2,10 @@
 
 #define PROJECT_NAME "RakeEngine"
 #define PROJECT_VERSION "0.0"
-#define LOCAL_BUILD_DATE "2024-04-02"
-#define LOCAL_BUILD_TIME "12:41:25"
-#define GMT_BUILD_DATE "2024-04-02"
-#define GMT_BUILD_TIME "11:41:25"
+#define LOCAL_BUILD_DATE "2024-04-04"
+#define LOCAL_BUILD_TIME "18:59:59"
+#define GMT_BUILD_DATE "2024-04-04"
+#define GMT_BUILD_TIME "17:59:59"
 
 // Metadata generated in local timezone: Etc/GMT-1
 // Good work fellow programmer!
