+++ Engine/STL/include/RKSTL/detection.h
