+++ Engine/STL/include/RKSTL/warnings.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "Detection.h"
+#include "detection.h"
 
 #define _SILENCE_CXX17_CODECVT_HEADER_DEPRECATION_WARNING
